+++ src/main/java/com/atheesh/app/ws/model/response/StoreResponse.java
@@ -11,7 +11,7 @@ public class StoreResponse {
     private Integer minLimit;
     private Integer unitQuantity;
     private String unitSymbol;
-    private Integer unitPrice;
+    private Float unitPrice;
     private String priceSymbol;
     private Status status;
 
@@ -78,11 +78,11 @@ public class StoreResponse {
         this.unitSymbol = unitSymbol;
     }
 
-    public Integer getUnitPrice() {
+    public Float getUnitPrice() {
         return unitPrice;
     }
 
-    public void setUnitPrice(Integer unitPrice) {
+    public void setUnitPrice(Float unitPrice) {
         this.unitPrice = unitPrice;
     }
