+++ src/main/java/com/atheesh/app/ws/repositories/StoreRepository.java
@@ -4,7 +4,6 @@ import com.atheesh.app.ws.entities.ItemEntity;
 import com.atheesh.app.ws.entities.ShopEntity;
 import com.atheesh.app.ws.entities.StoreEntity;
 import com.atheesh.app.ws.shared.enums.Status;
-import com.atheesh.app.ws.shared.enums.UserStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
@@ -25,5 +24,5 @@ public interface StoreRepository extends JpaRepository<StoreEntity, Integer> {
             "storeEntity.amount=:newAmount,storeEntity.minLimit=:newLimit,storeEntity.unitQuantity=:unitQuantity,storeEntity.unitSymbol=:unitSymbol,storeEntity.unitPrice=:newPrice,storeEntity.priceSymbol=:priceSymbol,storeEntity.status=:newStatus where storeEntity.id =:id")
     int updateTheStoreById(@Param("id") Integer id, @Param("newItem") ItemEntity newItem, @Param("newShop") ShopEntity newShop,
                           @Param("newAmount") Integer newAmount, @Param("newLimit") Integer newLimit,@Param("unitQuantity") Integer unitQuantity,
-                          @Param("unitSymbol") String unitSymbol,@Param("newPrice") Integer newPrice,@Param("priceSymbol") String priceSymbol,@Param("newStatus") Status newStatus);
+                          @Param("unitSymbol") String unitSymbol,@Param("newPrice") Float newPrice,@Param("priceSymbol") String priceSymbol,@Param("newStatus") Status newStatus);
 }
