+++ src/main/java/com/atheesh/app/ws/shared/dto/StoreDTO.java
@@ -11,14 +11,14 @@ public class StoreDTO {
     private Integer minLimit;
     private Integer unitQuantity;
     private String unitSymbol;
-    private Integer unitPrice;
+    private Float unitPrice;
     private String priceSymbol;
     private Status status;
 
     public StoreDTO() {
     }
 
-    public StoreDTO(Integer id, ItemDTO item, ShopDTO shop, Integer amount, Integer minLimit, Integer unitQuantity, String unitSymbol, Integer unitPrice, String priceSymbol, Status status) {
+    public StoreDTO(Integer id, ItemDTO item, ShopDTO shop, Integer amount, Integer minLimit, Integer unitQuantity, String unitSymbol, Float unitPrice, String priceSymbol, Status status) {
         this.id = id;
         this.item = item;
         this.shop = shop;
@@ -99,11 +99,11 @@ public class StoreDTO {
         this.unitSymbol = unitSymbol;
     }
 
-    public Integer getUnitPrice() {
+    public Float getUnitPrice() {
         return unitPrice;
     }
 
-    public void setUnitPrice(Integer unitPrice) {
+    public void setUnitPrice(Float unitPrice) {
         this.unitPrice = unitPrice;
     }
